requirements.txt
fastapi
uvicorn
python-dotenv
python-telegram-bot[job-queue]
google-generativeai
twilio
supabase
gTTS
edge-tts
websockets
pytz
requests
